Tweak existing python files to accommodate for changes in the database

+++ src/api/models.py
@@ -88,7 +88,7 @@ class Tile(models.Model):
         unique_together = ["x_coordinate", "y_coordinate"]
         ordering = ["x_coordinate", "y_coordinate"]
 
-    tile_id = models.IntegerField(primary_key=True, null=False, unique=True)
+    tile_id = models.BigIntegerField(primary_key=True, null=False, unique=True)
     x_coordinate = models.IntegerField(null=False)
     y_coordinate = models.IntegerField(null=False)
     REQUIRED_FIELDS = [tile_id, x_coordinate, y_coordinate]
